partial
拆分四叉树节点类
