+++ Assets/Quadtree Collider Detection/Tests/Editor/API Tests/List Test/ListTest.cs
@@ -6,6 +6,9 @@ using NUnit.Framework;
 [TestFixture]
 public class ListTest
 {
+    /// <summary>
+    /// 测试List和LLinkedList的foreach速度
+    /// </summary>
     [Test]
     public void ForeachSpeed()
     {
@@ -37,5 +40,15 @@ public class ListTest
         stopwatch.Stop();
 
         Debug.Log("linkedlist：" + stopwatch.ElapsedMilliseconds);
+    }
+
+    [Test]
+    public void RemoveCount()
+    {
+        List<int> list = new List<int> { 1, 2, 3 };
+
+        list.Remove(1);
+
+        Debug.Log(list.Count == 2 ? "List.Remove 会导致总数减少" : "List.Remove不会导致总数减少");
     }
 }
